# backend/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/warehouses": {
      "get": {
        "operationId": "WarehouseController_getAll",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Whether to include deactivated warehouses",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WarehouseResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "List all physical warehouses",
        "tags": ["warehouses"]
      },
      "post": {
        "operationId": "WarehouseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WarehouseResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new physical warehouse (Admin Only)",
        "tags": ["warehouses"]
      }
    },
    "/warehouses/stock/product/{productId}": {
      "get": {
        "operationId": "WarehouseController_getProductStocks",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WarehouseStockResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get stock distribution across all warehouses for a product",
        "tags": ["warehouses"]
      }
    },
    "/warehouses/{id}/stock": {
      "get": {
        "operationId": "WarehouseController_getWarehouseStocks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WarehouseStockResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get all product inventory stocks located in a warehouse",
        "tags": ["warehouses"]
      },
      "put": {
        "operationId": "WarehouseController_updateStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WarehouseStockResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update product stock in a warehouse and atomically sync totalStockCache (Admin Only)",
        "tags": ["warehouses"]
      }
    },
    "/warehouses/{id}": {
      "get": {
        "operationId": "WarehouseController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WarehouseResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get warehouse details by UUID",
        "tags": ["warehouses"]
      },
      "put": {
        "operationId": "WarehouseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WarehouseResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update warehouse details (Admin Only)",
        "tags": ["warehouses"]
      },
      "delete": {
        "operationId": "WarehouseController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Warehouse UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": ["object", "null"],
                          "default": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Deactivate warehouse (Admin Only)",
        "tags": ["warehouses"]
      }
    },
...
  },
  "info": {
    "title": "Hyundai E-Commerce API",
    "description": "Hyundai E-Commerce Platform API for Commercial Vehicles, B2B Dealer Quoting, Warehouse Management, and Multi-channel Payment Processing.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      },
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["success"]
      },
...
      "WarehouseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "nameVi": {
            "type": "string",
            "example": "Kho Tổng Hà Nội"
          },
          "nameEn": {
            "type": ["string", "null"],
            "example": "Hanoi Central Warehouse"
          },
          "streetAddress": {
            "type": "string",
            "example": "Lô CN-01, Khu Công Nghiệp Đài Tư, 386 Nguyễn Văn Linh"
          },
          "district": {
            "type": "string",
            "example": "Long Biên"
          },
          "city": {
            "type": "string",
            "example": "Hà Nội"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          }
        },
        "required": [
          "id",
          "nameVi",
          "streetAddress",
          "district",
          "city",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "StockProductItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "nameVi": {
            "type": "string",
            "example": "Máy phát điện Diesel Hyundai DHY65KSE"
          },
          "slug": {
            "type": "string",
            "example": "may-phat-dien-diesel-hyundai-dhy65kse"
          },
          "totalStockCache": {
            "type": "number",
            "example": 15
          }
        },
        "required": ["id", "nameVi", "slug", "totalStockCache"]
      },
      "StockWarehouseItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "nameVi": {
            "type": "string",
            "example": "Kho Tổng Hà Nội"
          },
          "city": {
            "type": "string",
            "example": "Hà Nội"
          }
        },
        "required": ["id", "nameVi", "city"]
      },
      "WarehouseStockResponseDto": {
        "type": "object",
        "properties": {
          "warehouseId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "productId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90"
          },
          "stock": {
            "type": "number",
            "example": 10
          },
          "minStockWarning": {
            "type": "number",
            "example": 2
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "product": {
            "$ref": "#/components/schemas/StockProductItemDto"
          },
          "warehouse": {
            "$ref": "#/components/schemas/StockWarehouseItemDto"
          }
        },
        "required": [
          "warehouseId",
          "productId",
          "stock",
          "minStockWarning",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateWarehouseDto": {
        "type": "object",
        "properties": {
          "nameVi": {
            "type": "string",
            "example": "Kho Tổng Hà Nội",
            "description": "Warehouse name in Vietnamese"
          },
          "nameEn": {
            "type": "string",
            "example": "Hanoi Central Warehouse",
            "description": "Warehouse name in English"
          },
          "streetAddress": {
            "type": "string",
            "example": "Lô CN-01, Khu Công Nghiệp Đài Tư, 386 Nguyễn Văn Linh",
            "description": "Street address"
          },
          "district": {
            "type": "string",
            "example": "Long Biên",
            "description": "District / County"
          },
          "city": {
            "type": "string",
            "example": "Hà Nội",
            "description": "City / Province"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "default": true,
            "description": "Active status"
          }
        },
        "required": ["nameVi", "streetAddress", "district", "city"]
      },
      "UpdateStockDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f",
            "description": "Product UUID"
          },
          "stock": {
            "type": "number",
            "example": 10,
            "description": "Physical stock quantity available in this warehouse",
            "minimum": 0
          },
          "minStockWarning": {
            "type": "number",
            "example": 2,
            "default": 2,
            "description": "Threshold quantity to trigger low-stock warning",
            "minimum": 0
          }
        },
        "required": ["productId", "stock"]
      },
      "UpdateWarehouseDto": {
        "type": "object",
        "properties": {
          "nameVi": {
            "type": "string",
            "example": "Kho Tổng Hà Nội Cập Nhật"
          },
          "nameEn": {
            "type": "string",
            "example": "Hanoi Main Warehouse Updated"
          },
          "streetAddress": {
            "type": "string",
            "example": "386 Nguyễn Văn Linh"
          },
          "district": {
            "type": "string",
            "example": "Long Biên"
          },
          "city": {
            "type": "string",
            "example": "Hà Nội"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}